exclude .md file from being deleted

+++ clear_bin.py
@@ -14,6 +14,8 @@ for root, dirs, files in os.walk(root_dir, topdown=False):
         if dir_name.startswith('chapter') or dir_name == 'playground':
             target_files = os.listdir(os.path.join(root_dir,dir_name))
             for target_file in target_files:
+                if target_file.endswith(".md"):
+                    continue
                 type_, encodeing = mimetypes.guess_type(target_file)
                 if type_ == None:
                     os.remove(
